EZP-26668: move Location Content fields to field mapper
