+++ lib/DocumentMapper/NativeDocumentMapper.php
@@ -17,15 +17,11 @@ use EzSystems\EzPlatformSolrSearchEngine\DocumentMapper;
 use eZ\Publish\SPI\Persistence\Content;
 use eZ\Publish\SPI\Persistence\Content\Type as ContentType;
 use eZ\Publish\SPI\Persistence\Content\Location;
-use eZ\Publish\SPI\Persistence\Content\Section;
 use eZ\Publish\SPI\Search\Field;
 use eZ\Publish\SPI\Search\Document;
 use eZ\Publish\SPI\Search\FieldType;
-use eZ\Publish\SPI\Persistence\Content\Handler as ContentHandler;
 use eZ\Publish\SPI\Persistence\Content\Location\Handler as LocationHandler;
 use eZ\Publish\SPI\Persistence\Content\Type\Handler as ContentTypeHandler;
-use eZ\Publish\SPI\Persistence\Content\ObjectState\Handler as ObjectStateHandler;
-use eZ\Publish\SPI\Persistence\Content\Section\Handler as SectionHandler;
 use eZ\Publish\Core\Search\Common\FieldRegistry;
 use eZ\Publish\Core\Search\Common\FieldNameGenerator;
 
@@ -67,13 +63,6 @@ class NativeDocumentMapper implements DocumentMapper
     protected $fieldRegistry;
 
     /**
-     * Content handler.
-     *
-     * @var \eZ\Publish\SPI\Persistence\Content\Handler
-     */
-    protected $contentHandler;
-
-    /**
      * Location handler.
      *
      * @var \eZ\Publish\SPI\Persistence\Content\Location\Handler
@@ -86,20 +75,6 @@ class NativeDocumentMapper implements DocumentMapper
      * @var \eZ\Publish\SPI\Persistence\Content\Type\Handler
      */
     protected $contentTypeHandler;
-
-    /**
-     * Object state handler.
-     *
-     * @var \eZ\Publish\SPI\Persistence\Content\ObjectState\Handler
-     */
-    protected $objectStateHandler;
-
-    /**
-     * Section handler.
-     *
-     * @var \eZ\Publish\SPI\Persistence\Content\Section\Handler
-     */
-    protected $sectionHandler;
 
     /**
      * Field name generator.
@@ -117,11 +92,8 @@ class NativeDocumentMapper implements DocumentMapper
      * @param \EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentTranslationFieldMapper $contentTranslationFieldMapper
      * @param \EzSystems\EzPlatformSolrSearchEngine\FieldMapper\LocationFieldMapper $locationFieldMapper
      * @param \eZ\Publish\Core\Search\Common\FieldRegistry $fieldRegistry
-     * @param \eZ\Publish\SPI\Persistence\Content\Handler $contentHandler
      * @param \eZ\Publish\SPI\Persistence\Content\Location\Handler $locationHandler
      * @param \eZ\Publish\SPI\Persistence\Content\Type\Handler $contentTypeHandler
-     * @param \eZ\Publish\SPI\Persistence\Content\ObjectState\Handler $objectStateHandler
-     * @param \eZ\Publish\SPI\Persistence\Content\Section\Handler $sectionHandler
      * @param \eZ\Publish\Core\Search\Common\FieldNameGenerator $fieldNameGenerator
      */
     public function __construct(
@@ -131,11 +103,8 @@ class NativeDocumentMapper implements DocumentMapper
         ContentTranslationFieldMapper $contentTranslationFieldMapper,
         LocationFieldMapper $locationFieldMapper,
         FieldRegistry $fieldRegistry,
-        ContentHandler $contentHandler,
         LocationHandler $locationHandler,
         ContentTypeHandler $contentTypeHandler,
-        ObjectStateHandler $objectStateHandler,
-        SectionHandler $sectionHandler,
         FieldNameGenerator $fieldNameGenerator
     ) {
         $this->blockFieldMapper = $blockFieldMapper;
@@ -144,11 +113,8 @@ class NativeDocumentMapper implements DocumentMapper
         $this->contentTranslationFieldMapper = $contentTranslationFieldMapper;
         $this->locationFieldMapper = $locationFieldMapper;
         $this->fieldRegistry = $fieldRegistry;
-        $this->contentHandler = $contentHandler;
         $this->locationHandler = $locationHandler;
         $this->contentTypeHandler = $contentTypeHandler;
-        $this->objectStateHandler = $objectStateHandler;
-        $this->sectionHandler = $sectionHandler;
         $this->fieldNameGenerator = $fieldNameGenerator;
     }
 
@@ -162,13 +128,7 @@ class NativeDocumentMapper implements DocumentMapper
     public function mapContentBlock(Content $content)
     {
         $locations = $this->locationHandler->loadLocationsByContent($content->versionInfo->contentInfo->id);
-        $section = $this->sectionHandler->load($content->versionInfo->contentInfo->sectionId);
         $mainLocation = null;
-        $locationFields = array();
-
-        foreach ($locations as $location) {
-            $locationFields[$location->id] = $this->mapLocationFields($location, $content, $section);
-        }
 
         $contentType = $this->contentTypeHandler->load($content->versionInfo->contentInfo->contentTypeId);
 
@@ -214,7 +174,6 @@ class NativeDocumentMapper implements DocumentMapper
                     array(
                         'id' => $this->generateLocationDocumentId($location->id, $languageCode),
                         'fields' => array_merge(
-                            $locationFields[$location->id],
                             $translationFields['regular'],
                             $metaFields,
                             $blockFields,
@@ -351,171 +310,6 @@ class NativeDocumentMapper implements DocumentMapper
         }
 
         return $fieldSets;
-    }
-
-    protected function mapLocationFields(Location $location, Content $content, Section $section)
-    {
-        $fields = [];
-
-        // UserGroups and Users are Content, but permissions cascade is achieved through
-        // Locations hierarchy. We index all ancestor Location Content ids of all
-        // Locations of an owner.
-        $ancestorLocationsContentIds = $this->getAncestorLocationsContentIds(
-            $content->versionInfo->contentInfo->ownerId
-        );
-        // Add owner user id as it can also be considered as user group.
-        $ancestorLocationsContentIds[] = $content->versionInfo->contentInfo->ownerId;
-        $fields[] = new Field(
-            'content_owner_user_group',
-            $ancestorLocationsContentIds,
-            new FieldType\MultipleIdentifierField()
-        );
-
-        $fields[] = new Field(
-            'content_id',
-            $content->versionInfo->contentInfo->id,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_type',
-            $content->versionInfo->contentInfo->contentTypeId,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_version_no',
-            $content->versionInfo->versionNo,
-            new FieldType\IntegerField()
-        );
-        $fields[] = new Field(
-            'content_status',
-            $content->versionInfo->status,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_name',
-            $content->versionInfo->contentInfo->name,
-            new FieldType\StringField()
-        );
-        $fields[] = new Field(
-            'content_creator',
-            $content->versionInfo->creatorId,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_owner',
-            $content->versionInfo->contentInfo->ownerId,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_section',
-            $content->versionInfo->contentInfo->sectionId,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_section_identifier',
-            $section->identifier,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_section_name',
-            $section->name,
-            new FieldType\StringField()
-        );
-        $fields[] = new Field(
-            'content_remote_id',
-            $content->versionInfo->contentInfo->remoteId,
-            new FieldType\IdentifierField()
-        );
-        $fields[] = new Field(
-            'content_modified',
-            $content->versionInfo->contentInfo->modificationDate,
-            new FieldType\DateField()
-        );
-        $fields[] = new Field(
-            'content_published',
-            $content->versionInfo->contentInfo->publicationDate,
-            new FieldType\DateField()
-        );
-        $fields[] = new Field(
-            'language_code',
-            array_keys($content->versionInfo->names),
-            new FieldType\MultipleStringField()
-        );
-        $fields[] = new Field(
-            'content_always_available',
-            $content->versionInfo->contentInfo->alwaysAvailable,
-            new FieldType\BooleanField()
-        );
-        $fields[] = new Field(
-            'content_group',
-            $this->contentTypeHandler->load(
-                $content->versionInfo->contentInfo->contentTypeId
-            )->groupIds,
-            new FieldType\MultipleIdentifierField()
-        );
-        $fields[] = new Field(
-            'content_object_state',
-            $this->getObjectStateIds($content->versionInfo->contentInfo->id),
-            new FieldType\MultipleIdentifierField()
-        );
-
-        return $fields;
-    }
-
-    /**
-     * Returns Content ids of all ancestor Locations of all Locations
-     * of a Content with given $contentId.
-     *
-     * Used to determine user groups of a user with $contentId.
-     *
-     * @param int|string $contentId
-     *
-     * @return array
-     */
-    protected function getAncestorLocationsContentIds($contentId)
-    {
-        $locations = $this->locationHandler->loadLocationsByContent($contentId);
-        $ancestorLocationContentIds = array();
-        $ancestorLocationIds = array();
-
-        foreach ($locations as $location) {
-            $locationIds = explode('/', trim($location->pathString, '/'));
-            // Remove Location of Content with $contentId
-            array_pop($locationIds);
-            // Remove Root Location id (id==1 in legacy DB)
-            array_shift($locationIds);
-
-            $ancestorLocationIds = array_merge($ancestorLocationIds, $locationIds);
-        }
-
-        foreach (array_unique($ancestorLocationIds) as $locationId) {
-            $location = $this->locationHandler->load($locationId);
-
-            $ancestorLocationContentIds[$location->contentId] = true;
-        }
-
-        return array_keys($ancestorLocationContentIds);
-    }
-
-    /**
-     * Returns an array of object state ids of a Content with given $contentId.
-     *
-     * @param int|string $contentId
-     *
-     * @return array
-     */
-    protected function getObjectStateIds($contentId)
-    {
-        $objectStateIds = array();
-
-        foreach ($this->objectStateHandler->loadAllGroups() as $objectStateGroup) {
-            $objectStateIds[] = $this->objectStateHandler->getContentState(
-                $contentId,
-                $objectStateGroup->id
-            )->id;
-        }
-
-        return $objectStateIds;
     }
 
     /**
