+++ lib/Resources/config/container/solr.yml
@@ -54,11 +54,8 @@ services:
             - "@ezpublish.search.solr.field_mapper.content_translation"
             - "@ezpublish.search.solr.field_mapper.location"
             - "@ezpublish.search.common.field_registry"
-            - "@ezpublish.spi.persistence.content_handler"
             - "@ezpublish.spi.persistence.location_handler"
             - "@ezpublish.spi.persistence.content_type_handler"
-            - "@ezpublish.spi.persistence.object_state_handler"
-            - "@ezpublish.spi.persistence.section_handler"
             - "@ezpublish.search.common.field_name_generator"
 
     ezpublish.search.solr.document_mapper:
