+++ lib/Resources/config/container/solr/field_mappers.yml
@@ -2,6 +2,7 @@ parameters:
     ezpublish.search.solr.field_mapper.content.content_document_base_fields.class: EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentFieldMapper\ContentDocumentBaseFields
     ezpublish.search.solr.field_mapper.content.content_document_location_fields.class: EzSystems\EzPlatformSolrSearchEngine\FieldMapper\ContentFieldMapper\ContentDocumentLocationFields
     ezpublish.search.solr.field_mapper.location.location_document_base_fields.class: EzSystems\EzPlatformSolrSearchEngine\FieldMapper\LocationFieldMapper\LocationDocumentBaseFields
+    ezpublish.search.solr.field_mapper.location.location_document_content_fields.class: EzSystems\EzPlatformSolrSearchEngine\FieldMapper\LocationFieldMapper\LocationDocumentContentFields
 
 services:
     ezpublish.search.solr.field_mapper.content.content_document_base_fields:
@@ -25,5 +26,16 @@ services:
         class: '%ezpublish.search.solr.field_mapper.location.location_document_base_fields.class%'
         arguments:
             - '@ezpublish.spi.persistence.content_handler'
+        tags:
+            - {name: ezpublish.search.solr.field_mapper.location}
+
+    ezpublish.search.solr.field_mapper.location.location_document_content_fields:
+        class: '%ezpublish.search.solr.field_mapper.location.location_document_content_fields.class%'
+        arguments:
+            - '@ezpublish.spi.persistence.content_handler'
+            - '@ezpublish.spi.persistence.location_handler'
+            - '@ezpublish.spi.persistence.content_type_handler'
+            - '@ezpublish.spi.persistence.object_state_handler'
+            - '@ezpublish.spi.persistence.section_handler'
         tags:
             - {name: ezpublish.search.solr.field_mapper.location}
